+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [1.1.7] - 2025-06-21
+### Fixes
+- Bug fix to `/colour` command
+- Bug fix to tracking of event posts
+
+## [1.1.6] - 2025-06-21
+### Added
+- Colour Role with the `/colour` command
+
 ## [1.1.5e] - 2025-06-21
 ### Added
 - Deleting the previous day's upcoming event schedules upon new day's message